organized project structure. This file system should make the project easier to test and debug
