+++ src/components/ui/Spinner.jsx
@@ -1,0 +1,11 @@
+import styles from "./Spinner.module.css";
+
+function Spinner() {
+  return (
+    <div className={styles.spinnerContainer}>
+      <div className={styles.spinner}></div>
+    </div>
+  );
+}
+
+export default Spinner;
