+++ src/layouts/Footer.jsx
@@ -1,0 +1,5 @@
+function Footer() {
+  return <footer></footer>;
+}
+
+export default Footer;
