+++ src/layouts/PageContainer.jsx
@@ -1,0 +1,5 @@
+function PageContainer({ children }) {
+  return <main>{children}</main>;
+}
+
+export default PageContainer;
